Move files to /source/; rename index to legacy
