+++ .gitignore
@@ -7,8 +7,4 @@ Thumbs.db
 *.bak
 *.log
 logs
-*.map
-/*.js
-/*.d.ts
-/*.test-d.js
-!/*.test.js
+distribution
