+++ source/cache-item.test-d.ts
@@ -1,5 +1,5 @@
-import {expectType, expectNotAssignable, expectNotType} from 'tsd';
-import {CacheItem} from './cache-item.js';
+import {expectType, expectNotAssignable, expectAssignable} from 'tsd';
+import CacheItem from './cache-item.js';
 
 type Primitive = boolean | number | string;
 type Value = Primitive | Primitive[] | Record<string, any>;
@@ -9,10 +9,10 @@ const item = new CacheItem<string>('key');
 expectType<Promise<boolean>>(item.isCached());
 expectType<Promise<void>>(item.delete());
 
-expectType<Promise<Value | undefined>>(item.get());
-expectType<Promise<string | undefined>>(item.get());
+expectAssignable<Promise<Value | undefined>>(item.get());
 expectNotAssignable<Promise<number | undefined>>(item.get());
-expectNotType<Promise<string>>(item.set('string'));
+expectType<Promise<string | undefined>>(item.get());
+expectType<Promise<string>>(item.set('some string'));
 
 // @ts-expect-error Type is string
 await item.set(1);
