+++ source/index.ts
@@ -1,0 +1,3 @@
+export {default as legacyCache} from './legacy.js';
+export {default as CacheItem} from './cache-item.js';
+export {default as UpdatableCacheItem} from './updatable-cache-item.js';
