+++ source/updatable-cache-item.test-d.ts
@@ -1,13 +1,13 @@
 /* eslint-disable no-new  */
 import {expectType, expectNotAssignable, expectNotType} from 'tsd';
-import {UpdatableCacheItem} from './updatable-cache-item.js';
+import UpdatableCacheItem from './updatable-cache-item.js';
 
 const itemWithUpdater = new UpdatableCacheItem('key', {
 	updater: async (one: number): Promise<string> => String(one).toUpperCase(),
 });
 
 expectType<((n: number) => Promise<string>)>(itemWithUpdater.get);
-expectNotAssignable<((n: number) => Promise<string>)>(itemWithUpdater.get);
+expectNotAssignable<((n: string) => Promise<string>)>(itemWithUpdater.get);
 
 async function identity(x: string): Promise<string>;
 async function identity(x: number): Promise<number>;
@@ -21,9 +21,8 @@ expectType<Promise<string>>(new UpdatableCacheItem('identity', {updater: identit
 // @ts-expect-error -- If a function returns undefined, it's not cacheable
 new UpdatableCacheItem('identity', {updater: async (n: undefined[]) => n[1]});
 
-// TODO: These expectation assertions are not working…
-expectNotAssignable<Promise<symbol>>(new UpdatableCacheItem('identity', {updater: identity}).get(1));
-expectNotType<Promise<symbol>>(new UpdatableCacheItem('identity', {updater: identity}).get('1'));
+expectNotAssignable<Promise<string>>(new UpdatableCacheItem('identity', {updater: identity}).get(1));
+expectNotType<Promise<number>>(new UpdatableCacheItem('identity', {updater: identity}).get('1'));
 
 new UpdatableCacheItem('number', {
 	updater: async (n: string) => Number(n),
