+++ vite.config.ts
@@ -4,7 +4,7 @@ import {defineConfig} from 'vitest/config';
 export default defineConfig({
 	test: {
 		setupFiles: [
-			'./test/_setup.js',
+			'./vitest.setup.js',
 		],
 	},
 });
